Add files via upload
This is my very first time using Figma. I followed a tutorial by Rino - LivingWithPixels in YouTube. The contents are completely different, but the layout is similar since it was my first time using Figma.

The design is for a cafe called 'Rightly Brewed' (fake cafe) where they introduce a new coffee in their menu.